refactor: 퀘스트 상수, 엔티티 리팩토링
