--- src/constants/date.ts
@@ -1,3 +1,0 @@
-export const START_YEAR = 2020
-
-export const MONTH = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11] as const
