+++ src/constants/window-view/repeat-quest-update-window-view.tsx
@@ -1,5 +1,5 @@
 import { WindowViewProps } from '@/components/window'
-import { RepeatQuest } from '@/entities'
+import { RepeatQuest } from '@/entities/quest'
 import { RepeatQuestEditPage } from '@/pages/quest/repeat-quest-edit-page'
 
 export const REPEAT_QUEST_UPDATE_PAGE_ID = (id: string) => `REPEAT_QUEST_UPDATE_${id}`
