+++ src/entities/index.ts
@@ -1,4 +1,3 @@
-export type * from './quest'
 export type * from './member'
 export type * from './history'
 export type * from './streak'
