--- src/entities/quests/index.ts
@@ -1,2 +1,0 @@
-export * from './quest-type'
-export * from './quest-status'
