+++ src/views/quest/components/quest-list-item.tsx
@@ -3,8 +3,7 @@ import { Checkbox } from '@/components/checkbox'
 import { useContextMenu } from '@/components/context-menu'
 import { IconButtons } from '@/components/icon-button'
 import { Iconify } from '@/components/iconify'
-import { QUEST_TYPE_LABEL } from '@/constants'
-import { AssignQuest } from '@/entities'
+import { AssignQuest, QUEST_TYPE } from '@/entities/quest'
 import { IContextMenuItem } from '@/stores/use-context-menu-store'
 import { useModalStore } from '@/stores/use-modal-store'
 import {
@@ -140,7 +139,7 @@ export function QuestListItem({
         <Stack direction="row" alignItems="center" gap={1}>
           <Box width={30} />
           <Stack direction="row" alignItems="center">
-            <Typography variant="caption">{QUEST_TYPE_LABEL[quest.questType]}퀘스트</Typography>
+            <Typography variant="caption">{QUEST_TYPE[quest.questType].label}</Typography>
             <Iconify icon="mdi:dot" sx={{ width: 15 }} width={15} />
             <Typography variant="caption">{quest.subjectName}</Typography>
             <Typography
