+++ src/views/quest/modals/create-quest/create-quest-modal.tsx
@@ -2,7 +2,7 @@ import { useAssignQuest } from '@/api/hooks'
 import { InvisibleContainer } from '@/components/container'
 import { ModalView } from '@/components/modal'
 import { Stepper } from '@/components/stepper'
-import { QuestType } from '@/entities'
+import { QuestType } from '@/entities/quest'
 import { Interest } from '@/entities/interest'
 import { useEnter } from '@/hooks'
 import { useModalStore } from '@/stores/use-modal-store'
